Add env setting to allow automatic registration of oauth users
New env setting: OAUTH_AUTO_REGISTER_ENABLED. Default is false which
means on a private instance any users authenticating with oauth will
be met with the not invited page. If the tenant is private and this
new setting is set to true then they are registered automatically
using the info in the oauth response.